+++ Coimbra.Services.ApplicationStateEvents/IApplicationStateEvent.cs
@@ -7,5 +7,6 @@ namespace Coimbra.Services.ApplicationStateEvents
     /// Base event interface for the Unity's application callbacks events.
     /// </summary>
     [RequireImplementors]
+    [AllowEventServiceUsageFor(typeof(IApplicationStateService))]
     public interface IApplicationStateEvent : IEvent { }
 }