+++ Coimbra.Services.PlayerLoopEvents/PlayerLoopSystem.cs
@@ -19,32 +19,10 @@ namespace Coimbra.Services.PlayerLoopEvents
         private IEventService? EventService => OwningLocator?.Get<IEventService>();
 
         /// <inheritdoc/>
-        public void RemoveAllListeners()
+        public void RemoveAllListeners<T>()
+            where T : IPlayerLoopEvent
         {
-            if (EventService == null)
-            {
-                return;
-            }
-
-            FixedUpdateEvent.RemoveAllListenersAt(EventService);
-            LateUpdateEvent.RemoveAllListenersAt(EventService);
-            UpdateEvent.RemoveAllListenersAt(EventService);
-            FirstEarlyUpdateEvent.RemoveAllListenersAt(EventService);
-            FirstFixedUpdateEvent.RemoveAllListenersAt(EventService);
-            FirstInitializationEvent.RemoveAllListenersAt(EventService);
-            FirstPostLateUpdateEvent.RemoveAllListenersAt(EventService);
-            FirstPreUpdateEvent.RemoveAllListenersAt(EventService);
-            FirstUpdateEvent.RemoveAllListenersAt(EventService);
-            LastEarlyUpdateEvent.RemoveAllListenersAt(EventService);
-            LastFixedUpdateEvent.RemoveAllListenersAt(EventService);
-            LastInitializationEvent.RemoveAllListenersAt(EventService);
-            LastPostLateUpdateEvent.RemoveAllListenersAt(EventService);
-            LastPreUpdateEvent.RemoveAllListenersAt(EventService);
-            LastUpdateEvent.RemoveAllListenersAt(EventService);
-            PostLateUpdateEvent.RemoveAllListenersAt(EventService);
-            PostTimeUpdateEvent.RemoveAllListenersAt(EventService);
-            PreLateUpdateEvent.RemoveAllListenersAt(EventService);
-            PreTimeUpdateEvent.RemoveAllListenersAt(EventService);
+            EventService?.RemoveAllListeners<T>();
         }
 
         /// <inheritdoc/>