+++ Roslyn~/Coimbra.Services.Events.Roslyn/CoimbraServicesEventsTypes.cs
@@ -6,6 +6,8 @@ namespace Coimbra.Services.Events.Roslyn
     {
         public const string Namespace = "Coimbra.Services.Events";
 
+        public static readonly TypeString AllowEventServiceUsageForAttribute = new("AllowEventServiceUsageForAttribute", Namespace);
+
         public static readonly TypeString EventInterface = new("IEvent", Namespace);
 
         public static readonly TypeString EventServiceInterface = new("IEventService", Namespace);