+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [10.3.3] - UNRELEASED
 
+- Added `AllowEventServiceUsageForAttribute` to allow using `IEventService` methods directly when inside specific types.
 - Removed analyzer `COIMBRA0008` due it being redundant. Using `BaseTypeRequiredAttribute` instead.
 
 ## [10.3.2] - 2022-05-26